backend/requirements.txt
flask==2.3.3
transformers==4.35.0
torch==2.8.0

sentence-transformers==2.2.2
PyPDF2==3.0.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
pandas==2.2.2

numpy==1.26.4
scikit-learn==1.3.2
python-dotenv==1.0.0
gunicorn==21.2.0